Add the ability to manage 32 and 64-bit keys/values
Previously, we were going to support a `redirect` parameter to control on
a per-key and value basis whether to allow or disable registry
redirection. If allowed, then the OS would redirect puppet to the 32-bit
view of the registry. However, since it was a parameter, we could not
manage both the 32 and 64-bit versions of the same key in the same
manifest, due to the paths not being unique. And composite namevars
doesn't help.

This commit removes the redirect parameter and instead allows the key or
value path to be prefixed by `32:`. If specified, the 32-bit view of the
registry is used. If not-specified (default), the 64-bit view of the
registry is used.